#70: Simplify ChangeEngagementStatusDialog props and add memoization
