+++ src/components/Dashboard/Profile/sections/VolunteerHeader/ChangeEngagementStatusDialog.tsx
@@ -6,105 +6,20 @@ import {
   DialogButtonGroup,
   LargePrimaryButton,
 } from "../VolunteerProfileDocumentSection/shared/DialogButtonGroup";
+import { de, enUS } from "date-fns/locale";
 import { VolunteerStateEngagementType } from "need4deed-sdk";
-import { Locale } from "date-fns";
-import { TFunction } from "i18next";
-import styled from "styled-components";
-
-const ModalContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: var(--dialog-gap);
-  padding: 0;
-  height: 100%;
-`;
-
-const ModalTitle = styled.h3`
-  font-size: var(--dialog-title-font-size);
-  font-weight: var(--font-weight-bold);
-  line-height: var(--dialog-title-line-height);
-  letter-spacing: var(--dialog-title-letter-spacing);
-  color: var(--color-blue-700);
-  margin: 0;
-`;
-
-const OptionsContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: var(--dialog-gap);
-`;
-
-const OptionItem = styled.div`
-  display: flex;
-  flex-direction: column;
-  gap: var(--dialog-option-gap);
-`;
-
-const RadioOption = styled.label`
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  gap: var(--dialog-option-gap);
-  cursor: pointer;
-
-  input[type="radio"] {
-    appearance: none;
-    width: var(--dialog-radio-size);
-    height: var(--dialog-radio-size);
-    border: var(--dialog-radio-border-width) solid var(--color-grey-400);
-    border-radius: var(--percent-50);
-    cursor: pointer;
-    position: relative;
-    flex-shrink: 0;
-
-    &:checked {
-      border-color: var(--color-blue-700);
-
-      &::after {
-        content: "";
-        position: absolute;
-        width: 12px;
-        height: 12px;
-        border-radius: var(--percent-50);
-        background-color: var(--color-blue-700);
-        top: var(--percent-50);
-        left: var(--percent-50);
-        transform: translate(-50%, -50%);
-      }
-    }
-  }
-`;
-
-const OptionLabel = styled.span`
-  font-size: var(--dialog-label-font-size);
-  font-weight: var(--font-weight-regular);
-  line-height: var(--dialog-label-line-height);
-  letter-spacing: var(--letter-spacing-tight);
-  color: var(--color-blue-700);
-  flex: 1;
-`;
-
-const OptionDescription = styled.p`
-  font-size: var(--dialog-description-font-size);
-  font-weight: var(--font-weight-regular);
-  line-height: var(--dialog-description-line-height);
-  letter-spacing: var(--letter-spacing-tight);
-  color: var(--color-grey-700);
-  margin: 0;
-`;
-
-const DateFieldContainer = styled.div`
-  margin-top: var(--spacing-8);
-`;
-
-const ENGAGEMENT_DESCRIPTION_KEYS: Record<VolunteerStateEngagementType, string> = {
-  [VolunteerStateEngagementType.NEW]: "new_description",
-  [VolunteerStateEngagementType.ACTIVE]: "active_description",
-  [VolunteerStateEngagementType.AVAILABLE]: "available_description",
-  [VolunteerStateEngagementType.TEMP_UNAVAILABLE]: "tempUnavailable_description",
-  [VolunteerStateEngagementType.INACTIVE]: "inactive_description",
-  [VolunteerStateEngagementType.UNRESPONSIVE]: "unresponsive_description",
-};
+import { useTranslation } from "react-i18next";
+import { createEngagementLabelMap, ENGAGEMENT_DESCRIPTION_KEYS } from "./constants";
+import {
+  DateFieldContainer,
+  ModalContainer,
+  ModalTitle,
+  OptionDescription,
+  OptionItem,
+  OptionLabel,
+  OptionsContainer,
+  RadioOption,
+} from "./dialogStyles";
 
 type Props = {
   isOpen: boolean;
@@ -114,11 +29,7 @@ type Props = {
   onStatusChange: (status: VolunteerStateEngagementType) => void;
   returnDate: Date | undefined;
   onReturnDateChange: (date: Date | undefined) => void;
-  initialReturnDate: Date | undefined;
-  originalStatus: VolunteerStateEngagementType;
-  engagementLabelMap: Record<VolunteerStateEngagementType, string>;
-  locale: Locale;
-  t: TFunction;
+  isSaveDisabled: boolean;
 };
 
 export const ChangeEngagementStatusDialog = ({
@@ -129,21 +40,16 @@ export const ChangeEngagementStatusDialog = ({
   onStatusChange,
   returnDate,
   onReturnDateChange,
-  initialReturnDate,
-  originalStatus,
-  engagementLabelMap,
-  locale,
-  t,
+  isSaveDisabled,
 }: Props) => {
+  const { t, i18n } = useTranslation();
+  const locale = i18n.language === "de" ? de : enUS;
+  const engagementLabelMap = createEngagementLabelMap(t);
+
   const getEngagementDescription = (status: VolunteerStateEngagementType): string => {
     const key = ENGAGEMENT_DESCRIPTION_KEYS[status];
     return t(`dashboard.volunteerProfile.volunteerHeader.engagementStatus_options.${key}`);
   };
-
-  const isSaveDisabled =
-    statusEngagement === originalStatus &&
-    (statusEngagement !== VolunteerStateEngagementType.TEMP_UNAVAILABLE ||
-      returnDate?.getTime() === initialReturnDate?.getTime());
 
   return (
     <Modal isOpen={isOpen} onClose={onClose}>
